+++ .eslintignore
@@ -1,2 +1,4 @@
 .eslintrc.js
 webpack.config.js
+webpack.config-builder.js
+dev-server.js
