+++ package.json
@@ -20,7 +20,7 @@
     "prettier-styles": "prettier --write src/**/*.scss",
     "prettier": "prettier --write src/**/*.js",
     "build": "webpack",
-    "start": "webpack serve",
+    "start": "node dev-server.js",
     "test": "echo \"Error: no test specified\" && exit 0"
   },
   "author": "Heartex, Inc.",
@@ -47,21 +47,17 @@
     "chroma-js": "^2.1.0",
     "css-loader": "^5.0.2",
     "css-minimizer-webpack-plugin": "^1.2.0",
+    "date-fns": "^2.18.0",
     "dotenv-webpack": "^6.0.0",
     "eslint": "^7.20.0",
-    "eslint-plugin-prettier": "^3.1.3",
     "eslint-plugin-react": "^7.22.0",
     "html-webpack-plugin": "^5.2.0",
     "husky": "^3.1.0",
-    "less": "^3.12",
-    "less-loader": "^6.1.0",
     "lint-staged": "^9.5.0",
-    "match-sorter": "^4.2.0",
     "mini-css-extract-plugin": "^1.3.8",
     "mobx": "^5.15.4",
     "mobx-react": "^6",
     "mobx-state-tree": "^3.16.0",
-    "namor": "^2.0.2",
     "nanoid": "^3.1.10",
     "node-fetch": "^2.6.1",
     "prettier": "^2.0.5",
@@ -71,12 +67,9 @@
     "react-hot-loader": "^4.12.20",
     "react-hotkeys-hook": "^2.4.0",
     "react-icons": "^3.11.0",
-    "react-is": "^16.8.0",
-    "react-table": "^7.0.0",
     "react-virtualized-auto-sizer": "^1.0.2",
     "react-window": "^1.8.6",
     "react-window-infinite-loader": "^1.0.5",
-    "sass": "^1.26.3",
     "source-map-loader": "^2.0.1",
     "stylus": "^0.54.8",
     "stylus-loader": "^5.0.0",
