+++ public/index.html
@@ -2,8 +2,6 @@
 <html lang="en">
   <head>
     <meta charset="utf-8" />
-    <link href='//fonts.googleapis.com/css?family=Source+Sans+Pro:300,400,600|Roboto Mono' rel='stylesheet' type='text/css' />
-    <link href='//fonts.googleapis.com/css?family=Dosis:500&text=LabelStudio' rel='stylesheet' type='text/css' />
     <link href="//fonts.googleapis.com/css2?family=Roboto:ital,wght@0,300;0,400;0,500;0,700;1,300;1,400;1,500;1,700&display=swap" rel="stylesheet">
 
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
@@ -22,8 +20,6 @@
     <link href="/static/lsf/css/main.css" rel="stylesheet">
     <script src="/static/lsf/js/main.js"></script>
 
-    <script src="/main.js"></script>
-    <link rel="stylesheet" href="/main.css">
     <!--
       Notice the use of %PUBLIC_URL% in the tags above.
       It will be replaced with the URL of the `public` folder during the build.
@@ -33,7 +29,7 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
       -->
-    <title>Heartex DataManager 2.0</title>
+    <title><%= htmlWebpackPlugin.options.title %></title>
 
     <style>
       body {
@@ -79,6 +75,6 @@
       You need to enable JavaScript to run this app.
     </noscript>
 
-    <div id="app" data-project-id="9"></div>
+    <div id="app" data-project-id="11"></div>
   </body>
 </html>
